requirements.txt
click==8.2.1
colorama==0.4.6
dlib==20.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==2.2.6
opencv-python==4.11.0.86
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
